Add BadPhysics
Absolutely non optimized O(n^2) Physics

Fix Later

+++ complementary/handMade.h
@@ -4,7 +4,7 @@
 std::random_device rd;
 std::mt19937 rng(rd());
 
-int getRandomFloat(int min, int max) {
+float getRandomFloat(float min, float max) {
     std::uniform_real_distribution<float> dist(min, max);
     return dist(rng);
 }
